atualizacao gerenciador de acesso + ajustes componente de upload + novos componentes
