+++ package.json
@@ -15,10 +15,12 @@
     "@fortawesome/free-regular-svg-icons": "^6.4.0",
     "@fortawesome/free-solid-svg-icons": "^6.4.0",
     "@fortawesome/vue-fontawesome": "^3.0.3",
+    "@kyvg/vue3-notification": "^2.9.0",
     "@vee-validate/i18n": "^4.8.6",
     "@vee-validate/rules": "^4.8.6",
     "axios": "^1.3.6",
     "bootstrap": "^5.2.3",
+    "dropzone": "^6.0.0-beta.2",
     "jquery": "^3.6.4",
     "lodash": "^4.17.21",
     "pinia": "^2.0.35",
@@ -27,8 +29,7 @@
     "vue": "^3.2.47",
     "vue-multiselect": "^3.0.0-beta.1",
     "vue-router": "^4.1.6",
-    "vue-trumbowyg": "^4.0.0",
-    "vue3-dropzone": "^2.0.1"
+    "vue-trumbowyg": "^4.0.0"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.2.0",
