+++ src/App.vue
@@ -3,5 +3,8 @@
 </script>
 
 <template>
-	<RouterView/>
+	<div>
+		<router-view/>
+		<notifications position="bottom right" :duration="5000" />
+	</div>
 </template>
