+++ src/components/STooltip/index.js
@@ -1,0 +1,1 @@
+export { default as STooltip } from './STooltip.vue';
