+++ src/components/SUploads/index.js
@@ -1,1 +1,2 @@
-export { default as SUploads } from './SUploads.vue';
+export { default as SUpload } from './SUpload.vue';
+export { default as SDropzone } from './SDropzone.vue';
