+++ src/components/index.js
@@ -1,9 +1,10 @@
-export { SImage } from './SImage'
-export { SButton, SSubmitButton } from './SButtons'
-export { SButtonGroup } from './SButtonGroup';
-export { StretchedCard } from './SCards'
-export { STabs, STabItem } from './STabs';
-export { SMultiselect } from './SMultiselect';
-export { STextArea } from './STextArea';
-export { SAccordion, SAccordionList } from './SAccordion';
-export { SUploads } from './SUploads';
+export * from './SImage';
+export * from './SButtons';
+export * from './SButtonGroup';
+export * from './SCards';
+export * from './STabs';
+export * from './SMultiselect';
+export * from './STextArea';
+export * from './SAccordion';
+export * from './SUploads';
+export * from './STooltip';
