+++ src/exceptions/CustomError.js
@@ -1,0 +1,8 @@
+class CustomError extends Error {
+    constructor(message) {
+        super(message);
+        this.name = this.constructor.name;
+    }
+}
+
+export default CustomError;
