+++ src/main.js
@@ -1,5 +1,6 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
+import Notifications from '@kyvg/vue3-notification'
 import App from './App.vue';
 import router from './router';
 
@@ -21,6 +22,7 @@ app.component('font-awesome-icon', FontAwesomeIcon)
 
 app.use(router)
 app.use(pinia)
+app.use(Notifications)
 
 app.mount('#app')
