+++ src/middleware/AuthMiddleware.js
@@ -1,7 +1,7 @@
 import GerenciarAcesso from "@/helpers/GerenciarAcesso";
 
 export default async function log({ next, to }) {
-    await GerenciarAcesso.check();
+    await GerenciarAcesso.validateAccess();
 
     return next();
 }
