+++ src/mixins/Form/AppForm.js
@@ -1,6 +1,7 @@
+import _ from 'lodash';
 import { configure } from 'vee-validate';
 import { localize } from '@vee-validate/i18n';
-import { api, baseURL } from '@/api';
+import { api } from '@/api';
 import pt_BR from '@vee-validate/i18n/dist/locale/pt_BR.json';
 
 export default {
@@ -10,6 +11,16 @@ export default {
         action: {
             type: String,
             required: true,
+        },
+        data: {
+            type: Object,
+            default: function _default() {
+                return {};
+            }
+        },
+        responsiveBreakpoint: {
+            type: Number,
+            default: 850
         }
     },
     data() {
@@ -17,6 +28,33 @@ export default {
             form: {},
             redirectTo: '',
             submiting: false,
+            mediaCollections: [],
+            dimensions: {
+                w: window.innerWidth,
+                h: window.innerHeight,
+            },
+            onSmallScreen: window.innerWidth < this.responsiveBreakpoint,
+        }
+    },
+    filters: {
+        date: function date(date) {
+            var format = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 'YYYY-MM-DD';
+
+            var date = (0, _moment2.default)(date);
+            return date.isValid() ? date.format(format) : "";
+        },
+        datetime: function datetime(_datetime) {
+            var format = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 'YYYY-MM-DD HH:mm:ss';
+
+            var date = (0, _moment2.default)(_datetime);
+            return date.isValid() ? date.format(format) : "";
+        },
+        time: function time(_time) {
+            var format = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 'HH:mm:ss';
+
+            // '2000-01-01' is here just because momentjs needs a date
+            var date = (0, _moment2.default)('2000-01-01 ' + _time);
+            return date.isValid() ? date.format(format) : "";
         }
     },
     beforeCreate() {
@@ -24,10 +62,31 @@ export default {
             generateMessage: localize({ pt_BR }),
         });
     },
+    created() {
+        if (!_.isEmpty(this.data)) {
+            this.form = this.data;
+        }
+        window.addEventListener('resize', this.onResize);
+    },
     methods: {
+        getPostData() {
+            if (this.mediaCollections) {
+                this.mediaCollections.forEach(collection => {
+                    if (this.form[collection]) {
+                        console.warn("MediaUploader warning: Media input must have a unique name, '" + collection + "' is already defined in regular inputs.");
+                    }
+        
+                    if (this.$refs[collection + '_uploader']) {
+                        this.form[collection] = this.$refs[collection + '_uploader'].getFiles();
+                    }
+                });
+            }
+      
+            return this.form;
+        },
 		onSubmit() {
             this.submiting = true;
-            api.post(this.action, this.form)
+            api.post(this.action, this.getPostData())
                 .then(response => {
                     this.onSuccess(response.data)
                 })
@@ -47,6 +106,13 @@ export default {
             this.submiting = false;
 
             console.error(data);
-        }
+        },
+        onResize() {
+            this.onSmallScreen = window.innerWidth < this.responsiveBreakpoint,
+            this.dimensions = {
+                w: window.innerWidth,
+                h: window.innerHeight,
+            };
+        },
     },
 }
